% Generated by roxygen2 (4.0.2): do not edit by hand
\name{aFunction}
\alias{aFunction}
\title{Multiplies two numbers and returns the result}
\usage{
aFunction(a, b)
}
\arguments{
\item{a}{A number I want to multiply by something}

\item{b}{A multiplying constant}
}
\value{
A number - the product of a and b
}
\description{
Multiplies two numbers and returns the result
}
\author{
Paul Newcombe
}

