% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea_pacchetto.R
\name{crea_pacchetto}
\alias{crea_pacchetto}
\title{crea_pacchetto}
\usage{
crea_pacchetto(x = "./")
}
\arguments{
\item{x}{percorso dove installare i file di template}
}
\value{
Nulla
}
\description{
Inizializza un progetto RStudio (gia esistente) come pacchetto \emph{R} con template di file e percorsi
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
