% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_package.R
\name{create_package}
\alias{create_package}
\title{create_package}
\usage{
create_package(x = "./")
}
\arguments{
\item{x}{root folder for the project}
}
\value{
None
}
\description{
Create folder structures and template files to render an empty \emph{RStudio} project an $R$ package project
}
\author{
Luca Valnegri, \email{l.valnegri@datamaps.co.uk}
}
