% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_function_sum.R
\name{example_function_sum}
\alias{example_function_sum}
\title{example_function_sum}
\usage{
example_function_sum(a = NULL, b = NULL)
}
\arguments{
\item{a}{Default = NULL}

\item{b}{Default = NULL}
}
\value{
number
}
\description{
Function that sums two numbers
}
\examples{
\dontrun{
library(rtemplate)
rtemplate(1,1)
}
}
\keyword{sum}
