% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbind.R
\name{fbind}
\alias{fbind}
\title{Bind two factors}
\usage{
fbind(a, b)
}
\arguments{
\item{a}{factor}

\item{b}{factor}
}
\value{
factory
}
\description{
Creat a new factor from two existing factors
}
\examples{
fbind(iris$Species[c(1,2,3)], iris$Species[c(1,2,3)])
}
