% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RTconvert.R
\name{RTconvert}
\alias{RTconvert}
\title{Interpret a RT source in vector of strings and generates R code}
\usage{
RTconvert(
  lines,
  add = c(),
  mark.lines = FALSE,
  filename = "",
  included = FALSE,
  includedirs = NULL
)
}
\description{
Interpret a RT source in vector of strings and generates R code
}
\examples{
RTconvert(c(
 "Hello",
 "<?R cat('World') ?>"
))
}
