% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{sum_two_nums}
\alias{sum_two_nums}
\title{Sum function}
\usage{
sum_two_nums(a, b)
}
\arguments{
\item{a}{The first number}

\item{b}{The second number}
}
\value{
The sum of these two numbers
}
\description{
This is a dummy function set up to test the package
}
