% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RTconvert}
\alias{RTconvert}
\title{Interpret a RT source in vector of strings and generates R code}
\usage{
RTconvert(lines, add = c())
}
\description{
Interpret a RT source in vector of strings and generates R code
}
\examples{
RTconvert(c(
 "Hello",
 "<?R cat('World') ?>"
))
}

