% Generated by roxygen2 (4.0.2): do not edit by hand
\name{RTscript}
\alias{RTscript}
\title{Main sript for usage of RTemplate as a command-line tool}
\usage{
RTscript()
}
\description{
You just call it, and it takes arguments from command-line and does stuff.
an example shell script would look like this
}
\details{
"#!/bin/bash"
"R -e "rtemplate:RTscript()" --args $@"
}

